# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "目前仅仅是一个demo",
        "title": "Rest API",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
        "/AddMenu": {
            "post": {
                "description": "新增菜单信息 请求主体: base64(str=aaaa) 成功输出null",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "新增菜单信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "str",
                        "name": "str",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":{},\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/DeleteMenu": {
            "post": {
                "description": "删除菜单信息 请求主体: base64(str=aaaa) 成功输出null",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "删除菜单信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "str",
                        "name": "str",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":{},\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/GetAllMenuInfo": {
            "post": {
                "description": "前台条件获取菜单信息 请求主体: base64(ParameterStr=aaaa\u0026PageSize=10\u0026CurrentPage=1) 成功输出[]Menu",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "前台条件获取菜单信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ParameterStr",
                        "name": "ParameterStr",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "PageSize",
                        "name": "PageSize",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "CurrentPage",
                        "name": "CurrentPage",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":{[]Menu},\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Menu\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/GetAllMenuViewInfo": {
            "post": {
                "description": "前台条件获取菜单信息 请求主体: base64(ParameterStr=aaaa\u0026PageSize=10\u0026CurrentPage=1) 成功输出[]Menu",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "前台条件获取菜单信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ParameterStr",
                        "name": "ParameterStr",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "PageSize",
                        "name": "PageSize",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "CurrentPage",
                        "name": "CurrentPage",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":{[]MenuView},\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/GetCascaderMenu": {
            "post": {
                "description": "根据用户ID获取用户菜单信息 请求主体: Null  成功输出[]MenuTree",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "根据用户ID获取用户菜单信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":{[]MenuTree},\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/GetMenuAllCount": {
            "post": {
                "description": "获取菜单总条数 请求主体: Null  成功输出 int",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "获取菜单总条数",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":{int},\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/GetMenuByID": {
            "post": {
                "description": "根据菜单ID获取菜单信息 请求主体: base64(ID=aaaa) 成功输出Menu",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "根据菜单ID获取菜单信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "ID",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":{Menu},\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/SetMenuButton": {
            "post": {
                "description": "设置菜单按钮 请求主体: base64(menuId=aaaa\u0026buttonStr=bbbbbb) 成功输出",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "设置菜单按钮",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "菜单ID",
                        "name": "menuId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "按钮ID，拼着接字符串",
                        "name": "buttonStr",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":{},\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/UpdateMenu": {
            "post": {
                "description": "修改菜单信息 请求主体: base64(str=aaaa) 成功输出null",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "修改菜单信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token",
                        "name": "Token",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "str",
                        "name": "str",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"Code\":1,\"Data\":{},\"Message\":\"\"} or {\"Code\":-1,\"Data\":{},\"Message\":\"错误提示\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    }
}